Give dtype when setting array attributes in DarkNoiseMeas.purge_bursts()
Most of the times, it works fine without explictly setting dtype. However, for very ill-formed data where all of the spotted peaks are part of a burst (i.e. for very noisy data), frame_idx ends up being an empty list, and so, creating a numpy array out of it will create, by default, an empty array whose dtype is equal to np.float64. This makes the FrameIDX setter arise an error, because it expects a np.int64 array (even it it's an empty one).

+++ src/core/DarkNoiseMeas.py
@@ -1035,9 +1035,9 @@ class DarkNoiseMeas(SiPMMeas):
                 del amplitude[j]
                 del frame_idx[j]                
 
-        purged_copy.TimeDelay   = np.array(timedelay)
-        purged_copy.Amplitude   = np.array(amplitude)
-        purged_copy.FrameIDX    = np.array(frame_idx)
+        purged_copy.TimeDelay   = np.array(timedelay, dtype=np.float64)
+        purged_copy.Amplitude   = np.array(amplitude, dtype=np.float64)
+        purged_copy.FrameIDX    = np.array(frame_idx, dtype=np.int64)
 
         return purged_copy
